+++ api/src/routes/index.js
@@ -17,5 +17,6 @@ router.use('/users', require('./users') /* #swagger.security = [{"BearerAuth": [
 router.use('/workflows', require('./workflows') /* #swagger.security = [{"BearerAuth": []}] */);
 router.use('/projects', require('./projects') /* #swagger.security = [{"BearerAuth": []}] */);
 router.use('/statistics', require('./statistics'));
+router.use('uploads', require('./uploads'));
 
 module.exports = router;
